Conflicts:
	ui/addscientistdialog.cpp
	ui/addscientistdialog.h
	ui/addscientistdialog.ui

+++ ui/computerscientists.cpp
@@ -117,22 +117,43 @@ void ComputerScientists::on_btnComputerSearch_clicked()
 
 void ComputerScientists::on_btnRemoveScientist_clicked()
 {
-    auto selectedIndexes = ui->tableScientists->selectionModel()->selection().indexes();
-    for(int i = 0; i < selectedIndexes.size(); i+=6){ // += 6 to skip duplicate rows
-        auto item = ui->tableScientists->item(selectedIndexes.at(i).row(), selectedIndexes.at(i).column());
-        scientistService->remove(scientistList[item->type()]);
+    if(messageBox_are_you_sure()){
+        auto selectedIndexes = ui->tableScientists->selectionModel()->selection().indexes();
+        for(int i = 0; i < selectedIndexes.size(); i+=6){ // += 6 to skip duplicate rows
+            auto item = ui->tableScientists->item(selectedIndexes.at(i).row(), selectedIndexes.at(i).column());
+            scientistService->remove(scientistList[item->type()]);
+        }
     }
+    else
+        return;
 
     refreshScientists(); //TODO: what if the remove came from a search?
     //loadScientistTable(scientistList);
 }
 
+bool ComputerScientists::messageBox_are_you_sure()
+{
+    QMessageBox msgBox;
+    msgBox.setText("Are you sure you want to remove this from the list?");
+    //msgBox.setInformativeText("");
+    msgBox.setStandardButtons(QMessageBox::Yes | QMessageBox::No);
+    msgBox.setDefaultButton(QMessageBox::Yes);
+    int ret = msgBox.exec();
+    if(ret == 0x00010000) //0x00010000 value of "No"
+        return false;
+    else
+        return true;
+}
+
+
 void ComputerScientists::on_btnRemoveComputer_clicked()
 {
-    auto selectedIndexes = ui->tableComputers->selectionModel()->selection().indexes();
-    for(int i = 0; i < selectedIndexes.size(); i+=6){ // += 6 to skip duplicate rows
-        auto item = ui->tableComputers->item(selectedIndexes.at(i).row(), selectedIndexes.at(i).column());
-        computerService->remove(computerList[item->type()]);
+    if(messageBox_are_you_sure()){
+        auto selectedIndexes = ui->tableComputers->selectionModel()->selection().indexes();
+        for(int i = 0; i < selectedIndexes.size(); i+=6){ // += 6 to skip duplicate rows
+            auto item = ui->tableComputers->item(selectedIndexes.at(i).row(), selectedIndexes.at(i).column());
+            computerService->remove(computerList[item->type()]);
+        }
     }
 
     refreshComputers(); //TODO: what if the remove came from a search?
